Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

It was late Jacob, so I decided to only put the copyright and set it to Gerrit with the WIP sign.

Thanks anyway! I'll implement the changes in the other languages and add the translators of their respective language(s).

Patch-set: 2