Update patch set 1
Patch Set 1:

Eddy, this particular patch cannot cause what you mentioned above, its probably related to the earlier patches.

The gist of the last few theme manager patches was to try and figure out if a theme provides assets at a density 1 below/above the device's if not at the exact same density, in which case we don't want to show the "assets" warning dialog.

Previously we use to check the density of the preview image, which isn't exactly a good check because 90% of themes drop it under the drawable/ (density-agnostic folder), and hence you'd see the assets warning for pretty much majority of the themes out there.

If you can try the below patchset and give feedback it'd be helpful.

http://review.cyanogenmod.org/#/c/54500/

Patch-set: 1
Reviewer: Danesh M <940@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0